# openapi.yaml
openapi: 3.0.3
info:
  title: pglet API
  version: 0.1.0
  description: PostgreSQL database browser API
paths:
...
  /api/tables/{table}:
    get:
      operationId: getTableColumns
      summary: Get table column definitions
      parameters:
        - name: table
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Column list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Column'
  /api/tables/{table}/rows:
    get:
      operationId: getTableRows
      summary: Paginated table data
      parameters:
        - name: table
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: sort_column
          in: query
          schema:
            type: string
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [ASC, DESC]
      responses:
        '200':
          description: Paginated rows
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableRowsResult'
  /api/tables/{table}/info:
    get:
      operationId: getTableInfo
      summary: Table size and row estimates
      parameters:
        - name: table
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Table info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableInfo'
  /api/tables/{table}/indexes:
    get:
      operationId: getTableIndexes
      summary: Table indexes
      parameters:
        - name: table
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Index list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TableIndex'
  /api/tables/{table}/constraints:
    get:
      operationId: getTableConstraints
      summary: Table constraints
      parameters:
        - name: table
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Constraint list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TableConstraint'
  /api/functions/{function}:
    get:
      operationId: getFunctionDefinition
      summary: Get function or procedure source code
      parameters:
        - name: function
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Function definition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FunctionDefinition'
...
  /api/settings:
    get:
      operationId: getSettings
      summary: Get all settings
      responses:
        '200':
          description: Settings map
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
    put:
      operationId: updateSettings
      summary: Update settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
...
    Column:
      type: object
      required: [name, type, nullable, position, is_primary_key]
      properties:
        name:
          type: string
        type:
          type: string
        nullable:
          type: boolean
        default_value:
          type: string
          nullable: true
        comment:
          type: string
        position:
          type: integer
        is_primary_key:
          type: boolean
    TableInfo:
      type: object
      required: [total_size, table_size, index_size, row_estimate]
      properties:
        total_size:
          type: string
        table_size:
          type: string
        index_size:
          type: string
        row_estimate:
          type: integer
          format: int64
    TableIndex:
      type: object
      required: [name, definition, is_unique, is_primary]
      properties:
        name:
          type: string
        definition:
          type: string
        is_unique:
          type: boolean
        is_primary:
          type: boolean
        columns:
          type: array
          items:
            type: string
    TableConstraint:
      type: object
      required: [name, type, definition]
      properties:
        name:
          type: string
        type:
          type: string
        definition:
          type: string
        columns:
          type: array
          items:
            type: string
    CellValue:
      type: string
      nullable: true
      x-go-type: '*string'
...
    TableRowsResult:
      type: object
      required: [columns, column_types, rows, total_count, page, page_size]
      properties:
        columns:
          type: array
          items:
            type: string
        column_types:
          type: array
          items:
            type: string
        rows:
          type: array
          items:
            type: array
            items:
              $ref: '#/components/schemas/CellValue'
        total_count:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
...
    FunctionDefinition:
      type: object
      required: [name, schema, definition, language, arguments, return_type, volatility, kind]
      properties:
        name:
          type: string
        schema:
          type: string
        definition:
          type: string
        language:
          type: string
        arguments:
          type: string
        return_type:
          type: string
        volatility:
          type: string
        kind:
          type: string
...